refactor(canvas): split canvasStore god file into focused slices; place into empty canvas at camera
Two changes:

- Empty-canvas placement: beginPlacement now skips the ghost picker when the
  canvas has no nodes and drops the panel at the current camera centre. The
  recommendations only make sense when there's existing content to place around.

- Slice split: the 1833-line canvasStore is broken into single-responsibility
  slices under stores/canvas/ (history, nodes, viewport, placement, navigation,
  selection, regions, arrange) plus shared types, per-instance ctx, and helpers.
  Same single Zustand store via the slices pattern, so cross-slice get() calls
  are unchanged. canvasStore.ts keeps the public surface (factory, singleton,
  registry, selectors, re-exported types), shrinking to 259 lines. No change for
  the 28 external importers.
